Update final-object.interface.ts
This update introduces a password feature for hubs within the application. When a hub with a password is opened, a modal prompts the user to enter the password. If the entered password is incorrect, the input is cleared, and no interactions with the app are allowed until the correct password is provided.

To enhance security, passwords will be hashed before storage using bcrypt, making it more challenging for unauthorized users to access sensitive information. This implementation addresses user requests for increased privacy while acknowledging that the password is stored in an easily accessible format within the .vha2 files